+++ packages/inline-choice/controller/src/__tests__/index.test.js
@@ -28,79 +28,52 @@ describe('Controller', () => {
 
     describe('result', () => {
       it('should return result mode => evaluate', () => {
-        controller.model(baseModel, {selectedChoice: "a"}, mergeObj(baseEnv, {mode: "evaluate"}))
-          .then(result => {
-            expect(result.result).toMatchObject(correctChoice);
-          })
+        return controller.model(baseModel, {selectedChoice: "a"}, mergeObj(baseEnv, {mode: "evaluate"}))
+          .then(result => expect(result.result).toMatchObject(correctChoice))
       });
 
       it('should not return result mode => gather', () => {
-        controller.model(baseModel, {selectedChoice: "a"}, mergeObj(baseEnv, {mode: "gather"}))
-          .then(result => {
-            expect(result.result).toBeUndefined();
-          })
+        return controller.model(baseModel, {selectedChoice: "a"}, mergeObj(baseEnv, {mode: "gather"}))
+          .then(result => expect(result.result).toBeUndefined())
       });
 
       it('should not return result mode => view', () => {
-        controller.model(baseModel, {selectedChoice: "a"}, mergeObj(baseEnv, {mode: "view"}))
-          .then(result => {
-            expect(result.result).toBeUndefined();
-          })
+        return controller.model(baseModel, {selectedChoice: "a"}, mergeObj(baseEnv, {mode: "view"}))
+          .then(result => expect(result.result).toBeUndefined())
       });
     });
 
     describe('disabled', () => {
       it('should TRUE when mode => evaluate', () => {
-        controller.model(baseModel, {}, mergeObj(baseEnv, {mode: "evaluate"}))
-          .then(result => {
-            expect(result).toHaveProperty('disabled', true);
-          })
+        return controller.model(baseModel, {}, mergeObj(baseEnv, {mode: "evaluate"}))
+          .then(result => expect(result).toHaveProperty('disabled', true))
       });
 
       it('should TRUE when mode => view', () => {
-        controller.model(baseModel, {}, mergeObj(baseEnv, {mode: "view"}))
-          .then(result => {
-            expect(result).toHaveProperty('disabled', true);
-          })
+        return controller.model(baseModel, {}, mergeObj(baseEnv, {mode: "view"}))
+          .then(result => expect(result).toHaveProperty('disabled', true))
       });
 
       it('should FALSE when mode => gather', () => {
-        controller.model(baseModel, {}, mergeObj(baseEnv, {mode: "gather"}))
-          .then(result => {
-            expect(result).toHaveProperty('disabled', false);
-          })
+        return controller.model(baseModel, {}, mergeObj(baseEnv, {mode: "gather"}))
+          .then(result => expect(result).toHaveProperty('disabled', false))
       });
     });
 
 
     describe('choices', () => {
-      it('should return en-US choices', () => {
-        controller.model(baseModel, {}, mergeObj(baseEnv, {locale: "en-US"}))
-          .then(result => {
-            expect(result.choices).toMatchObject(baseModel.choices);
-          })
-      });
 
       it('should not return en-ES choices', () => {
-        controller.model(baseModel, {}, mergeObj(baseEnv, {locale: "en-ES"}))
-          .then(result => {
-            console.log("result", result);
-            expect(result.choices).toMatchObject(baseModel.choices);
-          })
+        return controller.model(baseModel, {}, mergeObj(baseEnv, {locale: "en-ES"}))
+          .then(result => expect(result.choices).toMatchObject([{}]))
       });
     });
 
     describe('mode', () => {
       it('should return env mode', () => {
-        controller.model(baseModel, {}, mergeObj(baseEnv, {mode: "view"}))
-          .then(result => {
-            expect(result.mode).toMatchObject('view');
-          })
+        return controller.model(baseModel, {}, mergeObj(baseEnv, {mode: "view"}))
+          .then(result => expect(result.mode).toMatch('view'))
       });
-    });
-
-    describe('prompt', () => {
-
     });
 
   });