+++ packages/inline-choice/src/inline-choice.jsx
@@ -1,14 +1,11 @@
 import React from "react";
 import PropTypes from 'prop-types';
 import { withStyles } from 'material-ui/styles';
-import Input from 'material-ui/Input';
-import { MenuItem } from 'material-ui/Menu';
-import { FormControl, FormHelperText } from 'material-ui/Form';
-import Select from 'material-ui/Select';
-import {Correct, Incorrect, CorrectResponse} from '@pie-libs/icons';
+import { FormControl } from 'material-ui/Form';
 import { indicators } from '@pie-libs/render-ui';
+import Choices from "./choices";
 
-// const {Correct} = indicators;
+const {Correct, Incorrect} = indicators;
 
 const styles = theme => ({
   container: {
@@ -44,36 +41,20 @@ class InlineChoice extends React.Component {
 
     const { choices, classes, disabled, result} = this.props;
 
-    const items = choices.map(function (item, index) {
-      return (
-        <MenuItem key={index} value={item.value}>{item.label[0].value}</MenuItem>
-      );
-    });
-
     let renderFeedback = function (result) {
       let { feedback } = result[0];
       return (
-        <FormHelperText>{feedback && feedback[0].value}</FormHelperText>
-      )
+        (result[0].correct) ? <Correct feedback={feedback && feedback[0].value}/> : <Incorrect feedback={feedback && feedback[0].value}/>
+      );
     }
 
     return (
       <div className={classes.container}>
         {choices.length > 0 && <FormControl className={classes.formControl} disabled={disabled}>
-          <Select
-            value={this.state.selected}
-            onChange={this.handleChange}
-            input={<Input name="input-choice" id="input-choice" />}
-          >
-            {items}
-          </Select>
-          {(result) && renderFeedback(result)}
+          <Choices items={choices} value={this.state.selected} onChange={this.handleChange} />
         </FormControl>}
-        {(result) && <div className={classes.formControl}>
-          {(result[0].correct) ? <Correct iconSet="emoji"/> : <Incorrect iconSet="emoji"/>}
-        </div>}
+        {result && renderFeedback(result)}
       </div>
-
     );
   }
 }