+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.1"></a>
+## [0.3.1](https://github.com/pieelements/pie-elements/compare/@pie-libs/icons@0.3.0...@pie-libs/icons@0.3.1) (2018-02-17)
+
+
+
+
+**Note:** Version bump only for package @pie-libs/icons
+
 <a name="0.3.0"></a>
 # [0.3.0](https://github.com/pieelements/pie-elements/compare/@pie-libs/icons@0.2.3...@pie-libs/icons@0.3.0) (2018-01-31)